% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_captioner.R
\name{pretty_captioner}
\alias{pretty_captioner}
\title{A wrapper for the captioner function}
\usage{
pretty_captioner(label = NULL, caption = NULL, prefix = "Figure",
  sec_prefix = NULL, auto_space = TRUE, levels = 1, type = NULL,
  infix = ".", display = "full", inline = FALSE, reinit = FALSE,
  cap_fun_name = "pretty_cap.cap", ...)
}
\arguments{
\item{label}{Character string containing a unique object name.}

\item{caption}{Character string containing the object caption.}

\item{prefix}{Character string containing text to go before object number.
The default is \emph{Figure}.}

\item{sec_prefix}{Character string containing text to between prefix and object number.
The default is to add nothing.}

\item{auto_space}{Logical indicating whether or not a space should automatically be added
following the prefix.  Space is added by default.}

\item{levels}{Logical or number indicating whether or not you want hierarchical numbering,
and if so, how many levels.  Hierarchical numbering is turned off by default.}

\item{type}{Vector with same length as \code{levels} indicating whether figure numbering
should be numeric (\emph{n}), lowercase character (\emph{c}), or uppercase character (\emph{C}).
If unspecified, \code{captioner} will revert to all numeric values.}

\item{infix}{Character string containing text to go between figure numbers if hierarchical
numbering is on.  Default is \emph{.}}

\item{display}{Character string (logical) indicating what display mode you would like:
full (or f) displays all information, cite (or c) displays just the prefix and number,
and num (or n) displays just the number.}

\item{inline}{Logical indicating whether the function should display the prefix uncapitalised/capitalised.}

\item{reinit}{Logical indicating whether to reinitialise the underlying captioner function.}

\item{cap_fun_name}{A character string indicating the name to assign to the underlying captioner function.}

\item{...}{Pass additional arguements to the intialised \code{\link[captioner]{captioner}} function}
}
\value{
A character string containing the table/figure number, and optionally the caption.
}
\description{
A wrapper for the \code{\link[captioner]{captioner}} function. This provides captioning functionality via \code{\link[captioner]{captioner}}
in word, html, and pdf. The wrapper can be used without initialisation, adds a secondary prefix, and an optional inline arguement.
It works by intialising a second \code{\link[captioner]{captioner}} function which can then be called within the main function.
Best practise is to set reinit = TRUE on first using the funciton.
}
\details{
As a wrapper for the \code{\link[captioner]{captioner}} function \code{\link[prettypublisher]{pretty_captioner}} intialises a
\code{\link[captioner]{captioner}} function in the global enviroment. This is then used by pretty_captioner
to return a character string containing the prefix, sec_prefix, and object number with or without a caption.
The initial numbering is determined based on the order of caption creation. However, this order is modified based on the citations you use.
The first object to be cited will be moved to the beginning of the list, becoming object 1. Changing captioner parameters (prefix, sec_prefix,
auto_space, levels, type, and infix) requires the captioner function to be reinitialised by setting reinit to TRUE. The captioner function
can also be initialised without reference to an object by not supplying a label or caption argument.
For more details see \code{\link[captioner]{captioner}}.
}
\examples{
## First call to pretty_captioner intialising the captioner function
pretty_captioner('1', 'Example caption', reinit = TRUE)

## Second call
pretty_captioner('2', 'Example caption 2')

## Displaying the cite information only for use inline
pretty_captioner('2', display = 'c', inline = TRUE)

## Reference in text without a caption
pretty_captioner('3')

##Changing the prefix, adding a sec_prefix and reinitialising
pretty_captioner('1', 'Example caption 1', prefix = 'Table', sec_prefix = 'S', reinit = TRUE)

## Defining captioner function without a subsequent call to the captioning function
pretty_captioner(prefix = 'Table', sec_prefix = 'S', reinit = TRUE)
}
\seealso{
\code{\link[prettypublisher]{prettypublisher}} has a group of related functions which use various defaults: \code{\link[prettypublisher]{pretty_captioner}},
 \code{\link[prettypublisher]{pretty_figref}}, \code{\link[prettypublisher]{pretty_supfigref}}, \code{\link[prettypublisher]{pretty_tabref}},
 and \code{\link[prettypublisher]{pretty_suptabref}}.
}
