% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_ci.R
\name{pretty_ci}
\alias{pretty_ci}
\title{Format an estimate, lower and upper confidence intervals into a single character string}
\usage{
pretty_ci(est, lci, uci, sep = " to ", digits = 2)
}
\arguments{
\item{est}{A numeric or character vector of estimates}

\item{lci}{A numeric or character vector of lower confidence/credible intervals}

\item{uci}{A numeric or character vector of upper confidence/credible intervals}

\item{sep}{A character vector indicating the seperator used between the upper and
lower confidence/credible intervals. The default is ' to '.}

\item{digits}{Integer indicating the number of decimal places to be used.}
}
\value{
The estimate with formated upper, and lower confidence/credible intervals as a character vector
}
\description{
Take numeric/character vectors of the estimate, and upper/lower confidence/credible
intervals, round with trailing zeros and format into the required journal style
}
\examples{
## Formating a single confidence interval
pretty_ci(2, lci = 1, uci = 3)

## Formating vectors of confidence intervals
est <- c(0, 1, 100, 300, 21221, 403)
lci <- c(-123, -0.2, 50, 100, 12321, 200)
uci <- c(10, 2, 200, 400, 30122, 500)

pretty_ci(est, lci = lci, uci = uci, sep = '-', digits = 1)

## Use in a dplyr workflow
library(dplyr)

x <- data_frame(est = c(0,1), lci = c(0, 2), uci = c(1, 4))
x <- x \%>\% mutate(ci = est \%>\% pretty_ci(lci = lci, uci = uci, sep = ' by ', digits = 0))

}
\seealso{
pretty_round pretty_percentage
}
