% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_inline_ci.R
\name{pretty_inline_ci}
\alias{pretty_inline_ci}
\title{Reformat an estimate, with lower/upper confidence/credible intervals with an explanatory note}
\usage{
pretty_inline_ci(string, note = "95\% CI ")
}
\arguments{
\item{string}{A character vector as formated by \code{\link[prettypublisher]{pretty_ci}}.}

\item{note}{A character vector indicating the explanatory note to be used.}
}
\value{
An estimate with formated CI's for use inline within a document.
}
\description{
Takes a character vector as formated by \code{\link[prettypublisher]{pretty_ci}}
and adds a user specified explanatory note; to be used inline.
}
\examples{
## Use input from pretty_ci to generate a inline CI
pretty_inline_ci(pretty_ci(2,1,3))

## Change note i.e for 90\% credible intervals
pretty_inline_ci(pretty_ci(2,1,3), note = '90\% credible interval ')

## Handle vectorised input
est <- c(1,2,3,10)
lci <- c(0, -19, -209, -100)
uci <- c(100, 30, 40, 19)

formated_ci <- pretty_ci(est, lci, uci, sep = '-')
pretty_inline_ci(formated_ci)

}
\seealso{
pretty_ci
}
