% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_percentage.R
\name{pretty_percentage}
\alias{pretty_percentage}
\title{Calculate a percentage, and then format for publication}
\usage{
pretty_percentage(num, denom, digits = 2, percent_scaling = 100,
  as_per = TRUE, ...)
}
\arguments{
\item{num}{A numeric vector.}

\item{denom}{A numeric vector}

\item{digits}{Integer indicating the number of decimal places to be used.}

\item{percent_scaling}{A numeric vector indicating the scaling used,
by default this is 100 as this gives a percentage.}

\item{as_per}{Logical indicating if the percentage should be deliminated by a \% symbol.}

\item{...}{Pass additional arguements to \code{\link[base]{format}}.}
}
\value{
A character vector or string, with a percentage and
the numerator and denominator used to calculate it
}
\description{
Take numeric vectors of numerators, and demoninators, round them,
 format them with trailing zeros, and return as a character string along with the original numeric values.'
}
\examples{
## Return a percentage
pretty_percentage(10, 500)

## Return a vector of percentages
pretty_percentage(c(1,4,6, 19), 100, digits = 1)

## Return a proportion
pretty_percentage(c(1, 2, 4, 5, 2), c(100, 40, 50, 40, 30), as_per = FALSE)
}
\seealso{
pretty_round pretty_ci
}
