% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_suptabref.R
\name{pretty_suptabref}
\alias{pretty_suptabref}
\title{A wrapper for pretty_captioner with sensible defaults for captioning supplementary tables}
\usage{
pretty_suptabref(label = NULL, caption = NULL,
  prefix = "Supplementary Table", sec_prefix = "S", auto_space = TRUE,
  levels = 1, type = NULL, infix = ".", display = "full",
  inline = FALSE, reinit = FALSE, ...)
}
\arguments{
\item{label}{Character string containing a unique object name.}

\item{caption}{Character string containing the object caption.}

\item{prefix}{Character string containing text to go before object number.
The default is \emph{Figure}.}

\item{sec_prefix}{Character string containing text to between prefix and object number.
The default is to add nothing.}

\item{auto_space}{Logical indicating whether or not a space should automatically be added
following the prefix.  Space is added by default.}

\item{levels}{Logical or number indicating whether or not you want hierarchical numbering,
and if so, how many levels.  Hierarchical numbering is turned off by default.}

\item{type}{Vector with same length as \code{levels} indicating whether figure numbering
should be numeric (\emph{n}), lowercase character (\emph{c}), or uppercase character (\emph{C}).
If unspecified, \code{captioner} will revert to all numeric values.}

\item{infix}{Character string containing text to go between figure numbers if hierarchical
numbering is on.  Default is \emph{.}}

\item{display}{Character string (logical) indicating what display mode you would like:
full (or f) displays all information, cite (or c) displays just the prefix and number,
and num (or n) displays just the number.}

\item{inline}{Logical indicating whether the function should display the prefix uncapitalised/capitalised.}

\item{reinit}{Logical indicating whether to reinitialise the underlying captioner function.}

\item{...}{Pass additional arguements to the intialised \code{\link[captioner]{captioner}} function}
}
\value{
A character string containing the table/figure number, and optionally the caption.
}
\description{
A wrapper for pretty_captioner with sensible defaults for captioning supplementary tables
}
\details{
As a wrapper for the \code{\link[captioner]{captioner}} function \code{\link[prettypublisher]{pretty_captioner}} intialises a
\code{\link[captioner]{captioner}} function in the global enviroment. This is then used by pretty_captioner
to return a character string containing the prefix, sec_prefix, and object number with or without a caption.
The initial numbering is determined based on the order of caption creation. However, this order is modified based on the citations you use.
The first object to be cited will be moved to the beginning of the list, becoming object 1. Changing captioner parameters (prefix, sec_prefix,
auto_space, levels, type, and infix) requires the captioner function to be reinitialised by setting reinit to TRUE.
For more details see \code{\link[captioner]{captioner}}.
}
\examples{

## Generate a simple reference
pretty_suptabref('1' , 'Example', reinit = TRUE)

## cite in text
pretty_suptabref('1')

## Drop captilisation for use in line
pretty_suptabref('1', inline = TRUE)

## Add another table caption
pretty_suptabref('2' , 'Example 2')

}
\seealso{
pretty_captioner pretty_suptabref pretty_figref pretty_supfigref
}
