% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_table.R
\name{pretty_table}
\alias{pretty_table}
\title{Wrapper function for pretty tables}
\usage{
pretty_table(df, col_names = NULL, footer = NULL, cap_fun = NULL,
  label = NULL, caption = NULL, tab_fun = pander, ...)
}
\arguments{
\item{df}{A data frame to be converted to a markdown table.}

\item{col_names}{A character vector of replacement column names.}

\item{footer}{The desired footer as a character string.}

\item{cap_fun}{Caption function to wrap.}

\item{label}{A character string of the reference label for the table}

\item{caption}{A character string of the required table caption.}

\item{tab_fun}{Table function to wrap. Supported functions are \code{\link[pander]{pander}}
and \code{\link[knitr]{kable}}, with \code{\link[pander]{pander}} set as the default.}

\item{...}{Pass additional arguements to the wrapped table function.}
}
\value{
A markdown table
}
\description{
Wraps table functions to provide a single  user interface.
A footer can be added but this requies rows to be merged once the document
has been knit to word.
}
\examples{

## A simple table
pretty_table(iris[1:5, 1:5])

## Renaming columns
pretty_table(iris[1:5, 1:5], col_names = as.character(1:5))

## Adding a footer
pretty_table(iris[1:5, 1:5], footer = 'Example footer')

## Changing to a kable table
library(knitr)

pretty_table(iris[1:5, 1:5], tab_fun = kable)
}
