% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettypublisher.R
\docType{package}
\name{prettypublisher-package}
\alias{prettypublisher}
\alias{prettypublisher-package}
\title{prettypublisher: Formatting Functions For Publishable R Markdown Documents}
\description{
Provides a variety of general purpose functions that help in formatting R markdown documents ready for publication. Specifically focussed on improving the appearance of word documents, although all functions should also work for both pdf, and html.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.samabbott.co.uk/prettypublisher}
  \item \url{https://github.com/seabbs/prettypublisher}
  \item Report bugs at \url{https://github.com/seabbs/prettypublisher/issues}
}

}
\author{
\strong{Maintainer}: Sam Abbott Developer \email{contact@samabbott.co.uk}

}
\keyword{internal}
