% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_round.R
\name{pretty_round}
\alias{pretty_round}
\title{Round and format numeric vectors; standardising trailing zeros}
\usage{
pretty_round(num, digits = 2, ...)
}
\arguments{
\item{num}{A numeric vector.}

\item{digits}{Integer indicating the number of decimal places to be used.}

\item{...}{Pass additional arguements to \code{\link[base]{format}}.}
}
\value{
A rounded numeric vector formatted as a character vector with trialing zeros.
}
\description{
Round and format numeric vectors; standardising trailing zeros
}
\examples{
## Rounding a single number
pretty_round(2.1, digits = 1)

## Rounding with trailing whitespace
pretty_round(2.1, digits = 2)

##Rounding a vector
pretty_round(c(2.13, 2.1, 4.165, 8.2323242, 1), digits = 2)
}
\seealso{
\code{\link[base]{format}}, and \code{\link[base]{round}} for additional information
}
